refactor: move proper util methods into utils module
